requirements.txt
about-time==4.2.1
astroid==3.0.2
asttokens==2.4.1
Brotli==1.1.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distlib==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.47.0
future==0.18.3
grapheme==0.6.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multipledispatch==1.0.0
mutagen==1.47.0
nest-asyncio==1.5.8
nodeenv==1.8.0
numpy==1.24.4
opencv-python==4.5.5.64
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.20.0
Pygments==2.17.2
pylint==3.0.3
pyparsing==3.1.1
pypotrace==0.3
pytest==7.4.4
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.13
websockets==12.0
yt-dlp==2023.12.30
